# backend/src/main/java/com/mlb/mlbportal/controllers/StadiumController.java
package com.mlb.mlbportal.controllers;

import java.io.IOException;
import java.net.URI;
import java.util.List;

import org.springframework.data.domain.Page;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.servlet.support.ServletUriComponentsBuilder;

import com.mlb.mlbportal.dto.stadium.CreateStadiumRequest;
import com.mlb.mlbportal.dto.stadium.StadiumDTO;
import com.mlb.mlbportal.dto.stadium.StadiumInitDTO;
import com.mlb.mlbportal.models.others.PictureInfo;
import com.mlb.mlbportal.security.jwt.AuthResponse;
import com.mlb.mlbportal.security.jwt.AuthResponse.Status;
import com.mlb.mlbportal.services.StadiumService;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.AllArgsConstructor;

@Tag(name = "Stadiums", description = "Endpoints related to MLB stadiums")
@RestController
@RequestMapping("/api/v1/stadiums")
@AllArgsConstructor
public class StadiumController {
    private final StadiumService stadiumService;

    @Operation(summary = "Get all stadiums", description = "Returns a list of all MLB stadiums with basic information such as name, location, and capacity.")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "Successfully retrieved list of stadiums", content = @Content(mediaType = "application/json", schema = @Schema(implementation = StadiumInitDTO.class))),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = @Content(mediaType = "application/json"))
    })
    @GetMapping(produces = "application/json")
    public ResponseEntity<Page<StadiumInitDTO>> getAllStadiums(
            @RequestParam(defaultValue = "0") int page,
            @RequestParam(defaultValue = "10") int size) {

        Page<StadiumInitDTO> response = this.stadiumService.getAllStadiums(page, size);
        return ResponseEntity.ok(response);
    }

    @Operation(summary = "Get available stadiums", description = "Returns a list of all available stadiums. An available stadium is one that does not have any associated team.")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "Successfully retrieved list of stadiums", content = @Content(mediaType = "application/json", schema = @Schema(implementation = StadiumInitDTO.class))),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = @Content(mediaType = "application/json"))
    })
    @GetMapping(value = "/available", produces = "application/json")
    public ResponseEntity<Page<StadiumInitDTO>> getAllAvailableStadiums(
            @RequestParam(defaultValue = "0") int page,
            @RequestParam(defaultValue = "10") int size) {

        Page<StadiumInitDTO> response = this.stadiumService.getAllAvailableStadiums(page, size);
        return ResponseEntity.ok(response);
    }

    @Operation(summary = "Get stadium by name", description = "Returns basic information about a specific stadium identified by its name.")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "Successfully retrieved stadium info", content = @Content(mediaType = "application/json", schema = @Schema(implementation = StadiumInitDTO.class))),
            @ApiResponse(responseCode = "404", description = "Stadium not found", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = @Content(mediaType = "application/json"))
    })
    @GetMapping(value = "/{name}", produces = "application/json")
    public ResponseEntity<StadiumInitDTO> getStadiumByName(@PathVariable("name") String name) {
        return ResponseEntity.ok(this.stadiumService.findStadiumByName(name));
    }

    @Operation(summary = "Get stadium pictures", description = "Retrieves all pictures associated with a specific stadium identified by its name.")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "Successfully retrieved stadium pictures", content = @Content(mediaType = "application/json", schema = @Schema(implementation = PictureInfo.class))),
            @ApiResponse(responseCode = "404", description = "Stadium not found or no pictures available", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = @Content(mediaType = "application/json"))
    })
    @GetMapping(value = "/{stadiumName}/pictures", produces = "application/json")
    public ResponseEntity<List<PictureInfo>> getStadiumPictures(@PathVariable("stadiumName") String stadiumName) {
        return ResponseEntity.ok(this.stadiumService.getStadiumPictures(stadiumName));
    }

    @Operation(summary = "Upload stadium picture", description = "Uploads a new picture for a specific stadium.")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "Picture uploaded successfully", content = @Content(mediaType = "application/json", schema = @Schema(implementation = PictureInfo.class))),
            @ApiResponse(responseCode = "400", description = "Invalid file format or request parameters", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "401", description = "Unauthorized access", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "404", description = "Stadium not found", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = @Content(mediaType = "application/json"))
    })
    @PostMapping(value = "/{stadiumName}/pictures", consumes = "multipart/form-data", produces = "application/json")
    public ResponseEntity<PictureInfo> uploadPicture(@PathVariable("stadiumName") String stadiumName,
            @RequestParam("file") MultipartFile picturePath) throws IOException {
        return ResponseEntity.ok(this.stadiumService.addPicture(stadiumName, picturePath));
    }

    @Operation(summary = "Update stadium map picture", description = "Updates the stadium map picture for a specific stadium.")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "Picture uploaded successfully", content = @Content(mediaType = "application/json", schema = @Schema(implementation = PictureInfo.class))),
            @ApiResponse(responseCode = "400", description = "Invalid file format or request parameters", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "401", description = "Unauthorized access", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "404", description = "Stadium not found", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = @Content(mediaType = "application/json"))
    })
    @PostMapping(value = "/{stadiumName}/pictures/map", consumes = "multipart/form-data", produces = "application/json")
    public ResponseEntity<PictureInfo> updateStadiumMapPicture(@PathVariable("stadiumName")String stadiumName, @RequestParam("file") MultipartFile file) throws IOException {
        return ResponseEntity.ok(this.stadiumService.editStadiumMapPicture(stadiumName, file));
    }

    @Operation(summary = "Delete stadium picture", description = "Deletes a picture from a specific stadium using its public identifier.")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "Picture deleted successfully", content = @Content(mediaType = "application/json", schema = @Schema(implementation = AuthResponse.class))),
            @ApiResponse(responseCode = "401", description = "Unauthorized access", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "404", description = "Stadium or picture not found", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = @Content(mediaType = "application/json"))
    })
    @DeleteMapping(value = "/{stadiumName}/pictures", produces = "application/json")
    public ResponseEntity<AuthResponse> deletePicture(@PathVariable("stadiumName") String stadiumName,
            @RequestParam("publicId") String publicId) {
        this.stadiumService.deletePicture(stadiumName, publicId);
        return ResponseEntity.ok(new AuthResponse(Status.SUCCESS, "Picture deleted"));
    }

    @Operation(summary = "Create a new stadium", description = "Creates a new stadium with the provided name and opening date. If the stadium already exists, a conflict error will be returned.")
    @ApiResponses(value = {
            @ApiResponse(responseCode = "201", description = "Stadium created successfully", content = @Content(mediaType = "application/json", schema = @Schema(implementation = StadiumDTO.class))),
            @ApiResponse(responseCode = "400", description = "Invalid request body or missing required fields", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "401", description = "Unauthorized access", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "409", description = "Stadium already exists", content = @Content(mediaType = "application/json")),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = @Content(mediaType = "application/json"))
    })
    @PostMapping(consumes = "application/json", produces = "application/json")
    public ResponseEntity<StadiumDTO> createStadium(@Valid @RequestBody CreateStadiumRequest request) {
        StadiumDTO newStadium = this.stadiumService.createStadium(request);
        URI location = ServletUriComponentsBuilder
                .fromCurrentRequest()
                .path("/{name}")
                .buildAndExpand(newStadium.name())
                .toUri();
        return ResponseEntity.created(location).body(newStadium);
    }
}