# docs/api/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
...
paths:
...
  /api/v1/stadiums:
    get:
      tags:
      - Stadiums
      summary: Get all stadiums
      description: "Returns a list of all MLB stadiums with basic information such\
        \ as name, location, and capacity."
      operationId: getAllStadiums
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved list of stadiums
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StadiumInitDTO"
    post:
      tags:
      - Stadiums
      summary: Create a new stadium
      description: "Creates a new stadium with the provided name and opening date.\
        \ If the stadium already exists, a conflict error will be returned."
      operationId: createStadium
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStadiumRequest"
        required: true
      responses:
        "400":
          description: Invalid request body or missing required fields
          content:
            application/json: {}
        "409":
          description: Stadium already exists
          content:
            application/json: {}
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "201":
          description: Stadium created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StadiumDTO"
  /api/v1/stadiums/{stadiumName}/pictures:
    get:
      tags:
      - Stadiums
      summary: Get stadium pictures
      description: Retrieves all pictures associated with a specific stadium identified
        by its name.
      operationId: getStadiumPictures
      parameters:
      - name: stadiumName
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Stadium not found or no pictures available
          content:
            application/json: {}
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved stadium pictures
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PictureInfo"
    post:
      tags:
      - Stadiums
      summary: Upload stadium picture
      description: Uploads a new picture for a specific stadium.
      operationId: uploadPicture
      parameters:
      - name: stadiumName
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "400":
          description: Invalid file format or request parameters
          content:
            application/json: {}
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Stadium not found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Picture uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PictureInfo"
    delete:
      tags:
      - Stadiums
      summary: Delete stadium picture
      description: Deletes a picture from a specific stadium using its public identifier.
      operationId: deletePicture
      parameters:
      - name: stadiumName
        in: path
        required: true
        schema:
          type: string
      - name: publicId
        in: query
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Stadium or picture not found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Picture deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/v1/stadiums/{stadiumName}/pictures/map:
    post:
      tags:
      - Stadiums
      summary: Update stadium map picture
      description: Updates the stadium map picture for a specific stadium.
      operationId: updateStadiumMapPicture
      parameters:
      - name: stadiumName
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "400":
          description: Invalid file format or request parameters
          content:
            application/json: {}
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Stadium not found
          content:
            application/json: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Picture uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PictureInfo"
  /api/v1/players/{playerName}/pictures:
    post:
      tags:
      - Players
      summary: Upload player picture
      description: Uploads or updates the profile picture of a specific player.
      operationId: uploadPicture_1
      parameters:
      - name: playerName
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "400":
          description: Invalid file format
          content: {}
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Player not found
          content: {}
        "401":
          description: Unauthorized access
          content:
            application/json: {}
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Picture successfully uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PictureInfo"
...
  /api/v1/stadiums/{name}:
    get:
      tags:
      - Stadiums
      summary: Get stadium by name
      description: Returns basic information about a specific stadium identified by
        its name.
      operationId: getStadiumByName
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Stadium not found
          content:
            application/json: {}
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved stadium info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StadiumInitDTO"
  /api/v1/stadiums/available:
    get:
      tags:
      - Stadiums
      summary: Get available stadiums
      description: Returns a list of all available stadiums. An available stadium
        is one that does not have any associated team.
      operationId: getAllAvailableStadiums
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "500":
          description: Internal server error
          content:
            application/json: {}
        "422":
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "200":
          description: Successfully retrieved list of stadiums
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StadiumInitDTO"
...
components:
  schemas:
...
    PictureInfo:
      type: object
      properties:
        url:
          type: string
        publicId:
          type: string
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          enum:
          - SUCCESS
          - FAILURE
        message:
          type: string
        error:
          type: string
...
    CreateStadiumRequest:
      type: object
      properties:
        name:
          type: string
        openingDate:
          type: integer
          format: int32
      required:
      - name
      - openingDate
    StadiumDTO:
      type: object
      properties:
        name:
          type: string
        openingDate:
          type: integer
          format: int32
        pictures:
          type: array
          items:
            $ref: "#/components/schemas/PictureInfo"
...
    StadiumInitDTO:
      type: object
      properties:
        name:
          type: string
        openingDate:
          type: integer
          format: int32
        teamName:
          type: string
        pictures:
          type: array
          items:
            $ref: "#/components/schemas/PictureInfo"
        pictureMap:
          $ref: "#/components/schemas/PictureInfo"
...